% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acplot3d.event.R
\name{acplot3d.event}
\alias{acplot3d.event}
\title{Interactive scatterplot of the (3 dimensional) points either plotted by color and size or number of points in each voxel corresponding to the value of the points.}
\usage{
acplot3d.event(
  imgout = NULL,
  segout = FALSE,
  names_img = "frame",
  fmt = "png",
  info = "",
  ndigits = NULL,
  vessel.pos = FALSE,
  vessel.pos.col = "blue",
  vessel.pos.size = 0.3,
  force = FALSE,
  cores = 1,
  windowRect = c(0, 49, 1680, 1028),
  cex.3d = 3,
  event = 1,
  t = 1,
  turns = 10,
  cruise = 2009116,
  TVG = TRUE,
  TVG.exp = 2,
  dir.data = NULL,
  Paout = TRUE,
  exact = FALSE,
  bgns = TRUE,
  pdns = TRUE,
  nrns = TRUE,
  hins = TRUE,
  kern = NULL,
  segpar = NULL,
  pamkpar = list(),
  nsind = 0.75,
  hins_add = 10,
  pdns_scale = 1e-14,
  TOV = 0,
  cal = 1,
  bmmd = NULL,
  breaks = 40,
  col = "combined",
  colpar = list(start = 0, end = 0.8, flip = TRUE),
  clamp = c(0, 1),
  shrink = TRUE,
  null.value = NA,
  white = 1,
  log = TRUE,
  color.bar = "y--",
  color.bar.lwd = 8,
  color.bar.adj = 0,
  color.bar.tadj = 0.1,
  color.bar.noWhite = TRUE,
  color.bar.nticks = 8,
  color.bar.tickw = 0.005,
  color.bar.tickcol = "black",
  db = TRUE,
  voxels.out = FALSE,
  esnm = "MS70",
  var = "vbsc",
  ind = list(-(1:150), NULL),
  range = list(),
  subset = NULL,
  ideal = TRUE,
  seabed = -12000,
  rot = 2,
  compensation = c("pitch", "roll"),
  plot = TRUE,
  cs.xyzlim = "g",
  add = FALSE,
  beamstypes = 1,
  size = cplot3d.size("pow", y = 20, par = 4),
  bottom = FALSE,
  bottom.res = 5,
  bottom.smooth = NULL,
  bottom.col = "topo.col",
  bottom.N = 1e+05,
  bottom.seg = list(sgth = 0.2, misM = 0.001, bwGp = 2, turns = 100),
  adds = NULL,
  xlim = NULL,
  ylim = NULL,
  zlim = NULL,
  view = c("free", "top", "bottom", "south", "west", "north", "east"),
  zoom = 0.7,
  fov = 60,
  school = FALSE,
  schoolcol = "purple",
  schoolsize = 0.3,
  schoolsample = 0.01,
  schoollen = 4,
  schoollwd = 1,
  schooltype = "p",
  schoolnumt = NULL,
  schoolcrop = FALSE,
  plot.seg = FALSE,
  seg.col = "green",
  seg.alpha = 0.2,
  subdivide = 3,
  excl.neg = TRUE,
  object = c("ellipsoid", "cuboid"),
  par = double(3),
  center = c(0, 0, 0),
  angle = 0,
  aspect = "iso",
  nticks = 5,
  origin = 1,
  xlab = "x",
  ylab = "y",
  zlab = "z",
  full.box = FALSE,
  edge.vpos = "",
  line.vpos = 0,
  at.vpos = NULL,
  cex.vpos = 1,
  col.vpos = "blue",
  clock = "bbl",
  cex.clock = 1,
  adj.clock = c(0, 0),
  format.clock = "Ping: indt\\nyyyy-mm-dd\\nHH:MM:SS.FFF",
  digits.clock = 2,
  lsp.clock = 0.04,
  col.clock = 4,
  sonar.grid = "frame",
  sonar.grid.col = c("orange", "cornflowerblue"),
  sonar.grid.lwd = 1,
  cs.pos = "g",
  cs.view = "g",
  sides = c("tb", "tb"),
  dens = c(200, 100, 100, 1),
  every = Inf,
  global.grid = FALSE,
  global.grid.lwd = 0.5,
  global.grid.lty = 1,
  ...
)
}
\arguments{
\item{imgout}{specifies the directory in which to put the image files of the 3D representation of the regenerated points (plotted using the package rgl). If imgout==FALSE, no image files are saved, if imgout==TRUE or imgout==NULL image files are stored in the default directory named "..../sv2pos/acca___/fmt", and if 'imgout' is a character string, a directory named accordingly is created and image files stored. If already existing, the used is asked to confirm overwriting the existing files.}

\item{segout}{specifies the directory in which to put the segmentation data files. If segout==FALSE, no image files are saved, if segout==TRUE or segout==NULL segmentation files are stored in the default directory named "..../sv2pos/acca___/tsd", and if 'segout' is a character string, a directory named accordingly is created and image files stored. If already existing, the used is asked to confirm overwriting the existing files.}

\item{names_img}{is a string representing the names of the png files (excluding numbering and the file extension ".png"). As an example names_img="frame" and 100<length(t)<1000 result in the names "frame001.png", "frame002.png", ... .}

\item{fmt}{is a string representing the format of the image files. Currently supported are "png", "ps", "eps", "tex", "pdf", "svg" and "pgf". If fmt==NULL no images will be saved, only plotted.}

\item{info}{is a string to be added to the frame and tsd names.}

\item{ndigits}{is the number of digits in the numbering indexes in the image names and .school file names. If ndigits=5 and 100<length(t)<1000 the resulting names will be "frame00001.png", "frame00002.png", ..., for the default values of 'fmt' and 'names_img'.}

\item{vessel.pos}{is TRUE if every elapsed vessel position is to be plotted by a single dot on the surface.}

\item{vessel.pos.col}{is the color of the vessel positions.}

\item{vessel.pos.size}{is the size of the vessel positions.}

\item{force}{is FALSE if the user should be aske for permission to overwrite files.}

\item{cex.3d}{is the character expansion of the labels and axes in 3D.}

\item{full.box}{is FALSE (default) to make the bounding box in front of the data invisible, and TRUE ot make the full bounding box visible.}
}
\description{
Interactive scatterplot of the (3 dimensional) points either plotted by color and size or number of points in each voxel corresponding to the value of the points.
}
\examples{
\dontrun{}

}
