% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acplot3d.event_default.R
\name{acplot3d.event_default}
\alias{acplot3d.event_default}
\title{Generate a series of 3-D echograms with the function cplot3d.event(), which are save as png files and can be animated.}
\usage{
acplot3d.event_default(event, t = NULL, range = list(pszx = -80:-10),
  size = "SU90", breaks = seq(-86, -40, 2), white = 5, acca = 5e-04,
  info = "every1_cs_v", bgns = FALSE, pdns = FALSE, hins = FALSE,
  hins_add = 50, cores = 1)
}
\arguments{
\item{event}{is the identifier of the event, either given as the number of the event, a string contained in the name of the event, or the path of the event directory.}

\item{beams}{optional list of beam configuration data in the tSD format.}

\item{type}{the type of conversion, either "Sv" or "TS".}

\item{linear}{logical: if TRUE (default) return the linear values (volume backscattering coefficient), if FALSE return dB values (mean volume backscattering).}

\item{use}{string vector naming the conversions to include. "conv" applies the power to Sv or TS conversion and "tvg" applies the time varied gain (range compensation).}

\item{inv}{logica: if TRUE do the inverse conversion from Sv or TS to power.}

\item{cali}{a list of calibration data as returned from readcalfile(). Must contain either 'gain' or 'sacr' or both.}

\item{Rlist}{StoX R list.}
}
\value{
A list of relevant data from the plot.
}
\description{
Generate a series of 3-D echograms with the function cplot3d.event(), which are save as png files and can be animated.
}
\examples{
\dontrun{
acplot3d.event_default(t=1:10) # Unfinished example}

}
