% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.clock.R
\name{add.clock}
\alias{add.clock}
\title{Add time and date to an rgl plot.}
\usage{
add.clock(clock = "bbl", utim = 0, indt = NULL, cex.clock = 1,
  adj.clock = c(0, 0),
  format.clock = "Ping: indt\\nyyyy-mm-dd\\nHH:MM:SS.FFF", digits.clock = 2,
  lsp.clock = 0.04, col.clock = 4, D = 3)
}
\arguments{
\item{clock}{is the position of the clock timing the frames, either given as a vector of length 3 representing the coordinates of the clock, as FALSE for no clock, or as one of}

\item{utim}{is the unix time to plot.}

\item{indt}{is the optional index number to add to the output.}

\item{cex.clock}{is the numeric character expansion value for the clock.}

\item{adj.clock}{is used in text3d() when plotting the date and time.}

\item{format.clock}{is the format of the date and time, specified as the input to utim2ftim(). Default is "yyyy-mm-dd\\nHH:MM:SS.FFF" resulting in date and time separated by a line break of width according to the value of 'lsp.clock'. All "\\n" apparing in 'format.clock' cause a line break.}

\item{digits.clock}{is a numeric specifying the number of digits of the seconds part of the time.}

\item{lsp.clock}{is the spacing factor for the line break, given as a multiple of the corresponding x-, y- or zlim. Default is 0.04.}

\item{col.clock}{is the color of the plotted time and date.}
}
\value{

}
\description{
Add time and date to an rgl plot.
}
\examples{
\dontrun{}

}
