% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.sonar.grid.R
\name{add.sonar.grid}
\alias{add.sonar.grid}
\title{Plots or calculates the limits of sonar grids for the MS70 or the ME70.}
\usage{
add.sonar.grid(
  data,
  xyzlim = NULL,
  xyzlim.out = FALSE,
  sonar.grid = "frame",
  sonar.grid.col = c("orange", "cornflowerblue"),
  sonar.grid.lwd = 1,
  cs.pos = "g",
  cs.view = "g",
  sides = c("tb", "tb"),
  dens = c(200, 100, 100, 1),
  every = Inf,
  ...
)
}
\arguments{
\item{data}{is a list containing the data to plot, as returned from read.event(..., var=c("vbsc", "voxels", "ctd", "beams", "vessel", "time")). May include school positions data$psxf, data$psyf and data$pszf.}

\item{xyzlim}{is a 2 x 3 matrix of the limits of the bounding frame.}

\item{xyzlim.out}{is TRUE if the xyzlim expanded to the limits of the sonar grid should be returned.}

\item{sonar.grid}{is a string representing whether to plot the sonar grid as a projection onto the surface ("proj"), as a grid around the sonar volume ("frame") or no grid ("").}

\item{sonar.grid.col}{is a vector of two strings specifying the color to use for plotting the sonar grid edges, where the first is the main color, and the second is the color to use for lines that are above the plotting frame, but that are plotted onto the top of the frame.}

\item{sonar.grid.lwd}{is the line width of the sonar grid.}

\item{cs.pos}{is a single character representing the coordinate system used for the plotted data (one of "g" for global or "v" for vessel).}

\item{cs.view}{is a single character representing the coordinate system used for the view point (one of "g" for global or "v" for vessel).}

\item{sides}{is a vector of two strings specifying on which sides of the sonar volume to plot edges of the beams and bows, respectively, where "t" is "top", "r" is "right", "b" is "bottom" and "l" is "left".}

\item{dens}{is a vector of 4 elements specifying}

\item{every}{(used in plot_volume.ME70()) is a single integer representing that the frame should be plotted every 'every' time step, where the first and last are included, or a vector of integers giving the time steps for which the frame should be plotted (not including the first and last time step).}

\item{...}{are inputs passed on to plot_volume.ME70() and plot_volume.MS70().}
}
\description{
Plots or calculates the limits of sonar grids for the MS70 or the ME70.
}
\examples{
\dontrun{}

}
