% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm.colorsx.R
\name{cm.colorsx}
\alias{cm.colorsx}
\title{Returns a color vector moving from cyan to magenta via white.}
\usage{
cm.colorsx(
  x,
  v = 1,
  start = 0,
  end = 1,
  clamp = NULL,
  breakpoint = 0.5,
  flip = FALSE,
  alpha = 1,
  ...
)
}
\arguments{
\item{x}{is a numeric object to which rainbow colors combined with lightening and darkening should be matched.}

\item{v}{is the value (lightness) of the colors.}

\item{start}{and 'end' are the start and end colors, given as values in [0,1].}

\item{breakpoint}{is the breakpoint between 0 and 1 where the color vector turns at white.}

\item{flip}{is true if the color scale should be reversed.}

\item{alpha}{is the transparency.}

\item{...}{variables passed on from other functions.}
}
\description{
Returns a color vector moving from cyan to magenta via white.
}
\examples{
\dontrun{
x=runif(1000)
pp(1,2)
plot(x, pch=".", cex=30, col=cm.colorsx(x))
plot(x, pch=".", cex=30, col=cm.colorsx(x,flip=TRUE))
}

}
