% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplot2d.TSD.R
\name{cplot2d.TSD}
\alias{cplot2d.TSD}
\title{2-D echogram.}
\usage{
cplot2d.TSD(data, t = "all", breaks = 40, col = "combined",
  colpar = list(start = 0, end = 0.8, flip = TRUE), null.value = NA,
  beamstypes = 1, grid = TRUE, adds = NULL, xlim = NULL, ylim = NULL,
  zlim = NULL, tlim = NULL, up = FALSE, freq = 1, wb = 1, rmar = 5,
  xaxis = c("time", "dist", "pings", "compr", "sparse"), gap = median,
  gapthr = 10, tol = 0.1, heave = c("interp", "pixel", "ignore"),
  x0 = NULL, unit = NULL, date = c("unique", "all", "none"),
  nticksx = 10, lwdfact = 55, region = NULL, cs.pos = "v", white = 0,
  log = TRUE, endcol = c("white", ""), esnm = "MS70", var = c("vbsc",
  "sgsc", "pr0s", "sgs0", "sgsE", "sgsi", "sgsI", "psis", "tlns"),
  ind = list(), range = list(), subset = NULL, plot = TRUE,
  clock = NULL, cex.clock = 1,
  format.clock = "Ping: indt\\nyyyy-mm-dd\\nHH:MM:SS.FFF", digits.clock = 2,
  col.clock = 4, ...)
}
\arguments{
\item{data}{is a list containing the data to plot. Must contain the sv field 'vbsc', or the Sv field 'mvbs', the sampling interval field 'sint' and the average speed of sound 'asps'.}

\item{breaks}{has two possible inputs: (1) the number of breaks of the scale on which the data 'z' are arranged, equally spaced between min(z) and max(z) (or logarithmically equaly spaced when log=TRUE). (2) a vector of values for the breaks given in dB values (volume backscattering strength Sv).}

\item{col}{is either the color vector if length(col)>1, or the color function to generate the colors from. Currently the color function name must be one of "rainbow", "grey", "heat.colors", "terrain.colors", "topo.colors" and "cm.colors", but other color functions may be inplemented in the future. Set color.bar=NULL to supress plotting the color bar.}

\item{colpar}{is a list of parameters used in colscale().}

\item{zlim}{is is the depth range (negative values).}

\item{tlim}{is the time range.}

\item{up}{should be set to TRUE for upwards oriented echosounders.}

\item{freq}{is the frequency to plot.}

\item{log}{is TRUE if log transformed data are to be plotted. If so, all values must be positive.}

\item{endcol}{is the color to use for values above the highest break value.}

\item{z}{is the z-position of the echosounder (negative when below the sea surface, 0 when ).}
}
\value{

}
\description{
2-D echogram.
}
\examples{
\dontrun{}

}
