% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplot3d.col_breaks.R
\name{cplot3d.col_breaks}
\alias{cplot3d.col_breaks}
\title{Internal: Generates colors and beaks for color 3d plots.}
\usage{
cplot3d.col_breaks(
  col,
  colpar,
  breaks,
  white,
  clamp = c(0, 1),
  log,
  rangevbsc,
  endcol
)
}
\description{
Internal: Generates colors and beaks for color 3d plots.
}
\examples{
\dontrun{}

}
