% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplot3d.plot.color.bar.R
\name{cplot3d.plot.color.bar}
\alias{cplot3d.plot.color.bar}
\title{Plots the color bar of a cplot3d-plot.}
\usage{
cplot3d.plot.color.bar(color.bar.col, breaks = 40, white = 1, log = TRUE,
  color.bar = "x--", color.bar.lwd = 8, color.bar.adj = 0,
  color.bar.tadj = 0.1, color.bar.noWhite = TRUE, color.bar.nticks = 8,
  color.bar.tickw = 0.005, color.bar.tickcol = "black", db = TRUE)
}
\arguments{
\item{color.bar.col}{is the color of the color bar, generated inside cplot3d.TSD().}

\item{breaks}{is the number of breaks of the scale on which the data 'z' are arranged (generated in cplot3d.TSD()).}

\item{white}{is an integer value representing the number of the lower breaks for which points should not be plotted (corresponding to white points).}

\item{log}{is TRUE if log transformed data are to be plotted. If so, all values must be positive.}

\item{color.bar}{is a string of three characters, where the first is one of "x", "y", "z", and denotes along which dimension the color bar should be parallel, the second and third characters are either "-" or "+", and denotes along which of the four edges parallel to the axis given by the first charater, the color bar should be put (where the axes are ranged "x", "y", "z"). color.bar="x-+" puts the color bar on the top south edge, if the x-axis is defined east, the y-axis north and the z-axis vertical.}

\item{color.bar.lwd}{is the line width of the color bar (thickness).}

\item{color.bar.adj}{is a vector of adjustment values for the color bar (in the x, y, and z direction), where the values are given as fractions of the size of the corresponding dimensions. If given as a single numeric, 'color.bar.adj' is interpreted as the adjustment in the lowest available dimension (for example in the y direciton if color.bar="x--").}

\item{color.bar.tadj}{is similar to 'color.bar.adj', but for the labels of the tickmarks on the color bar.}

\item{color.bar.noWhite}{is FALSE to include all values which are not plotted due to 'white' in the color bar.}

\item{color.bar.nticks}{is the desired number of ticks on the color bar, used in pretty().}

\item{color.bar.tickw}{is the thickness of the tick marks on the color bar, as a fraction of the plotting frame.}

\item{color.bar.tickcol}{is the color of the tich marks.}

\item{db}{is FALSE if the color bar should display linear (not desibel) values.}
}
\value{

}
\description{
Plots the color bar of a cplot3d-plot.
}
\examples{
\dontrun{}

}
