% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplot3d.size.R
\name{cplot3d.size}
\alias{cplot3d.size}
\title{Returns a function of one signel numeric, returning the sizes used when plotting voxels by cplot3d.TSD(). The input to the retunred function is the number of breaks of the color plot. The inputs to cplot3d.size() are the type of size function, the range of the size values and the parameters of the size function.}
\usage{
cplot3d.size(type = "lin", y = NULL, par = NULL)
}
\arguments{
\item{type}{is a three character string representing the type of function of the size. Current valid values are}

\item{y}{is a vector of the range of the size values.}

\item{par}{is a vector of 0, 1 or 2 elements representing the parameters to the size function.}
}
\value{

}
\description{
Returns a function of one signel numeric, returning the sizes used when plotting voxels by cplot3d.TSD(). The input to the retunred function is the number of breaks of the color plot. The inputs to cplot3d.size() are the type of size function, the range of the size values and the parameters of the size function.
}
\examples{
\dontrun{}

}
