% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplot3d.TSD.R
\name{cplot3d.TSD}
\alias{cplot3d.TSD}
\title{Interactive scatterplot of the (3 dimensional) position data and response data included in the input list 'data', where the position data must have names ("psxx", "psyx", "pszx"), ("x", "y", "z"), ("s1", "s2", "s3"), ("psx", "psy", "psz"), or be named "coords", including all postition data in a column matrix. The response data must be named "vbsc", "mvbs", "data", "z", "sv" or "b" (case insentive). If none of the position data names are recognized, the first element of 'data' is taken to be the response and the next three to be the postition data, if more than 3 elements are present.}
\usage{
cplot3d.TSD(
  data,
  breaks = 40,
  col = "combined",
  colpar = list(start = 0, end = 0.8, flip = TRUE),
  size = cplot3d.size("pow", y = 20, par = 4),
  clamp = c(0, 1),
  shrink = TRUE,
  null.value = NA,
  beamstypes = 1,
  white = 1,
  log = TRUE,
  color.bar = "x--",
  color.bar.lwd = 8,
  color.bar.adj = 0,
  color.bar.tadj = 0.1,
  color.bar.noWhite = TRUE,
  color.bar.nticks = 8,
  color.bar.tickw = 0.005,
  color.bar.tickcol = "black",
  db = TRUE,
  voxels.out = FALSE,
  endcol = c("", ""),
  esnm = "MS70",
  var = c("vbsc", "sgsc", "pr0s", "sgs0", "sgsE", "sgsi", "sgsI", "psis", "tlns"),
  ind = list(-(1:150), NULL),
  range = list(),
  subset = NULL,
  ideal = TRUE,
  seabed = -12000,
  rot = 2,
  compensation = c("pitch", "roll"),
  plot = TRUE,
  cs.xyzlim = "g",
  add = FALSE,
  adds = NULL,
  xlim = NULL,
  ylim = NULL,
  zlim = NULL,
  view = c("free", "top", "bottom", "south", "west", "north", "east"),
  zoom = 0.7,
  fov = 60,
  school = FALSE,
  schoolcol = "purple",
  schoolsize = 0.3,
  schoolsample = 0.01,
  schoollen = 4,
  schoollwd = 1,
  schooltype = "p",
  schoolcrop = FALSE,
  plot.seg = FALSE,
  seg.col = "green",
  seg.alpha = 0.2,
  subdivide = 3,
  excl.neg = TRUE,
  object = c("ellipsoid", "cuboid"),
  par = double(3),
  center = c(0, 0, 0),
  angle = 0,
  aspect = "iso",
  nticks = 5,
  origin = 1,
  xlab = "x",
  ylab = "y",
  zlab = "z",
  full.box = FALSE,
  edge.vpos = "",
  line.vpos = 0,
  at.vpos = NULL,
  cex.vpos = 1,
  col.vpos = "blue",
  clock = "bbl",
  cex.clock = 1,
  adj.clock = c(0, 0),
  format.clock = "Ping: indt\\nyyyy-mm-dd\\nHH:MM:SS.FFF",
  digits.clock = 2,
  lsp.clock = 0.04,
  col.clock = 4,
  sonar.grid = "frame",
  sonar.grid.col = c("orange", "cornflowerblue"),
  sonar.grid.lwd = 1,
  cs.pos = "g",
  cs.view = "g",
  sides = c("tb", "tb"),
  dens = c(200, 100, 100, 1),
  every = Inf,
  global.grid = FALSE,
  global.grid.lwd = 0.5,
  global.grid.lty = 1,
  ...
)
}
\arguments{
\item{data}{is a list containing the data to plot, as returned from read.event(...,var=c("vbsc","voxels","ctd","beams","vessel","time")). May include school positions data$psxf, data$psyf and data$pszf. If non-school voxel probabilities are given in an array 'pr0s', or thresholded segmentation values 'sgsc'/'sgs0' are given, these can be plotted using the input 'var'.}
}
\description{
Interactive scatterplot of the (3 dimensional) position data and response data included in the input list 'data', where the position data must have names ("psxx", "psyx", "pszx"), ("x", "y", "z"), ("s1", "s2", "s3"), ("psx", "psy", "psz"), or be named "coords", including all postition data in a column matrix. The response data must be named "vbsc", "mvbs", "data", "z", "sv" or "b" (case insentive). If none of the position data names are recognized, the first element of 'data' is taken to be the response and the next three to be the postition data, if more than 3 elements are present.
}
\examples{
\dontrun{}

}
