% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpplot3d.bottom.TSD.R
\name{cpplot3d.bottom.TSD}
\alias{cpplot3d.bottom.TSD}
\title{Adding bottom surface from segmented bottom voxels.}
\usage{
cpplot3d.bottom.TSD(data, bottom.res = 2, bottom.smooth = NULL,
  col = "topo.col", step.size = 10000, insteps = TRUE, ...)
}
\arguments{
\item{data}{is the list of required input data, specifically, the bottom segmentation masks (sgbt); the lengths of the beams (lenb); the number of beams (optional, numb); the voxel positions in the x and y direction (psxx and psyx, respectively); and the bottom points in the x, y, and z direction, generated by pplot3d.TSD() (psxr, psyr, and pszr, respectively).}

\item{bottom.res}{is a two element vector of bottom.res distances in the x and y direction respectively. Grid points are positioned so that the end points of the range of the data in each direction is kept, and the distance between grid points is approximately equal to the values given in 'bottom.res'. If 'bottom.res' is given as a single value, it is recycled to both directions.}

\item{bottom.smooth}{is the width of the Gaussian kernel by which to smooth the bottom surface.}

\item{col}{is either a list of parameters used in colscale(), such as x, fun=c("rainbow","grey","heat.colors","terrain.colors","topo.colors","cm.colors","combined"), start=0, end=0.8, h=1/6, s=1, v=1, sc=c(1,0.3), vc=c(0.7,1), flip=FALSE, bias=1, space=c("rgb","Lab"), interpolate=c("linear","spline","fadewhite","fadeblack"), breakpoint=NULL, alpha=1.}

\item{step.size}{is the number of points used to generate the bottom at each step, set low to avoid memory limitations slowing down the generation of bottom points.}
}
\value{

}
\description{
Adding bottom surface from segmented bottom voxels.
}
\examples{
\dontrun{}

}
