% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpplot3d.extract_var.R
\name{cpplot3d.extract_var}
\alias{cpplot3d.extract_var}
\title{Extracts the variable (acoustic or segmentation) specified by 'var' from the 'data'. Only one time step must be given in 'data'.}
\usage{
cpplot3d.extract_var(data, var)
}
\arguments{
\item{data}{is a list containing the data to plot, as returned from read.event(...,var=c("vbsc","voxels","ctd","beams","vessel","time")). May include school positions data$psxf, data$psyf and data$pszf. If non-school voxel probabilities are given in an array 'pr0s', or thresholded segmentation values 'sgsc'/'sgs0' are given, these can be plotted using the input 'var'.}

\item{var}{is a string or a vector for two strings, specifying the variable to extract. If var[2] starts with "+" or "-", and the rest of the string is the four character name of a segmentation variable, the variable specified by var[1] is extracted but only ("+") or everything but ("-") the data in the segmentation mask. Adding a "-" at the end of var[2], so that this has 6 characters expands the subtraction to everythinb past the first segmentation index of the beam.}
}
\description{
Extracts the variable (acoustic or segmentation) specified by 'var' from the 'data'. Only one time step must be given in 'data'.
}
\examples{
\dontrun{}

}
