% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{cpplot3d}
\alias{cpplot3d}
\alias{cpplot3d-package}
\title{3D-plotting and segmentation of sonar data}
\description{
This R package provides functions for creating 3D color and point plots using the rgl package, and segmentation of sonar data.
}
\details{
Details of the package
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/arnejohannesholmin/cpplot3d/issues}
}

}
\author{
\strong{Maintainer}: Arne Johannes Holmin \email{arnejh@imr.no}

}
