% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.rbeta02sgth.R
\name{echoIBM.rbeta02sgth}
\alias{echoIBM.rbeta02sgth}
\title{Returns segmentation threshold values 'c' corresponding to the input realtive lower schooling threshold values 'rbeta0'. The function only treats one time step, which implies that 'misM' must have length 1 if given (one school of one time step).}
\usage{
echoIBM.rbeta02sgth(rbeta0, misM = NULL, pow = 1, table.out = FALSE, ...)
}
\arguments{
\item{rbeta0}{is an array of relative lower schooling threshold values.}

\item{misM}{is the mean sv of the school at the time step, given if the input is given in lower schooling threshold (not relative lower schooling threshold).}

\item{pow}{is a value to raise the segmentation threshold values to the power of, which if pow<1 eases the strictness of the segmentations, suitable for bottom detection (pow=1/4 might be good):}

\item{table.out}{is TRUE to return only the table.}

\item{...}{are parameters used in approx().}
}
\description{
Returns segmentation threshold values 'c' corresponding to the input realtive lower schooling threshold values 'rbeta0'. The function only treats one time step, which implies that 'misM' must have length 1 if given (one school of one time step).
}
\examples{
\dontrun{}

}
