% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.segment.event_oneping_write.R
\name{echoIBM.segment.event_oneping_write}
\alias{echoIBM.segment.event_oneping_write}
\title{Writes one ping of segmented data from the Bayesian segmentation method.}
\usage{
echoIBM.segment.event_oneping_write(data, beams, time, sgsc, sgsE, ii, thisi,
  block, thisa, alpha, minsgsc, newfile, segfile, totallength, filesize,
  pr0s.out, dir.out, var, thisbeta0, thisbeta1, beta0, beta1, lst0, input_misM,
  misM, h, pow, startn, add, numt, factor, ...)
}
\arguments{
\item{data}{is the list of data including segmentation and acoustic data.}

\item{beams}{is a list of beam configuration data.}

\item{time}{is a list of time information.}

\item{sgsc}{is unbiased segmentatino mask.}

\item{sgsE}{is enlarged segmentatino mask.}

\item{ii}{is the current time step index.}

\item{thisi}{is the current time step.}

\item{thisa}{is the index of current alpha value.}

\item{minsgsc}{is the minimum number of segmented voxels required by meanSv.TSD().}

\item{newfile}{is locigal deciding wheter to create a new file due to the 'filesize' limit.}

\item{segfile}{is the segmentation file(s).}

\item{totallength}{is a vector of the sizes of the files.}

\item{pr0s.out}{is TRUE if the pobability of beta<beta0 should be returned.}

\item{dir.out}{is the directory in which to put the segmentaiton files.}

\item{Xcsz}{is the maximum file size.}
}
\value{

}
\description{
Writes one ping of segmented data from the Bayesian segmentation method.
}
\examples{
\dontrun{}

}
