% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.sesg2event.R
\name{echoIBM.sesg2event}
\alias{echoIBM.sesg2event}
\title{Creates a new event from segmentation files. The directory of the event must be created in advance, to ensure secure generation of the event.}
\usage{
echoIBM.sesg2event(
  event,
  pingsdir,
  t = 1,
  filesize = 3e+08,
  type = "scsg",
  segpar = NULL,
  ow = TRUE,
  cruise = 2009116,
  esnm = "MS70",
  dir.data = NULL
)
}
\arguments{
\item{event}{is the identifier of the event holding the segmentation files, either given as the number of the event, a string contained in the name of the event, or the path of the event directory.}

\item{pingsdir}{is the path to the new event.}

\item{t}{is the time points to process.}

\item{filesize}{is the maximum size of the files.}

\item{type}{is one of "sgsc", representing estimated segmentation data, and "sgs0", representing theoretical segmentation data.}

\item{segpar}{is a list of elements named "bwGp", "lsth"/"rlst", "usth"/"rust", or "sgth"/"sgt0" specifying the parameters of the segmentation data to read.}

\item{ow}{is TRUE if the user wish to overwrite existing files.}

\item{cruise}{is either the idenfication number of the cruise, given as specified by the IMR (yyyynnn), or the path to the directory containing the event to be read.}

\item{esnm}{is the name of the acoustical instrument, given as a four character string. Currently implemented are "EK60", "ME70", "MS70" and "SH80"/"SX80"/"SH90"/"SX90" (may be given in lover case)}

\item{dir.data}{is the path to the directory in which the projects are stored, defaulted by the variable Acoustics_datasets_directory().}
}
\description{
Creates a new event from segmentation files. The directory of the event must be created in advance, to ensure secure generation of the event.
}
\examples{
\dontrun{}

}
