% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fade.colorsx.R
\name{fade.colorsx}
\alias{fade.colorsx}
\title{Returns a color vector interpolated between the colors specified by 'start' and 'end', possitbly fading via white of black.}
\usage{
fade.colorsx(x, start = "red", end = "violet", breakpoint = 0.5,
  bias = 1, space = c("rgb", "Lab"), interpolate = c("linear", "spline",
  "fadewhite", "fadeblack"), flip = FALSE, ...)
}
\arguments{
\item{x}{is a numeric object to which rainbow colors combined with lightening and darkening should be matched.}

\item{start}{and 'end' are the start and end colors.}

\item{breakpoint}{is the breakpoint between 0 and 1 where the color vector turns at white.}

\item{flip}{is true if the color scale should be reversed.}

\item{...}{methods passed on to sub-functions (not used, but allowing for unused arguments).}

\item{alpha}{is the transparency.}
}
\value{

}
\description{
Returns a color vector interpolated between the colors specified by 'start' and 'end', possitbly fading via white of black.
}
\examples{
\dontrun{
x=runif(300)
pp(1,2)
plot(x, pch=".", cex=30, col=fade.colorsx(x, start="orange", end="green"))
plot(x, pch=".", cex=30, col=fade.colorsx(x, start="orange", end="green", flip=TRUE))
}

}
