% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fplot3d.event.R
\name{fplot3d.event}
\alias{fplot3d.event}
\title{Plots fish as points or lines in 3-D, from data "psxf", "psyf", "pszf", "rtzf", "rtxf" located in the event.}
\usage{
fplot3d.event(event, t = 1, schoolcol = "red2", schoolsize = 0.3,
  schoolsample = 0.01, schoollen = 4, schoollwd = 1,
  schooltype = "p", schoolnumt = NULL, add = FALSE, aspect = "iso",
  ...)
}
\arguments{
\item{event}{is the identifier of the event, either given as the number of the event, a string contained in the name of the event, or the path of the event directory.}

\item{t}{is either a vector of the numbers of the pings to be returned, as listed from 1 to the number of pings in the event, or a vector of time points given as strings "yyyymmddHHMMSS.FFF" or "HHMMSS.FFF" from which the range of the time points to be read is extracted. If t=="all", all files are read and if t=="none" an empty list is returned.}

\item{schoolcol}{is the color of the points of the fish positions if school==TRUE or a string representing the directory containing .school files to be read for fish positions.}

\item{schoolsize}{is the size of the points of the fish positions if school==TRUE or a string representing the directory containing .school files to be read for fish positions.}

\item{schoolsample}{is the proportion of the fish positions given by the .school specified by 'school', that are to be sampled for plotting. Useful when the number of fish is large.}

\item{schoollen}{is the length of the fish when plotted as lines.}

\item{schoollwd}{is the width of the fish when plotted as lines.}

\item{schooltype}{is either "p" (default) for plotting points a the fish positions, "l" for plotting the fish as lines in the direcitons of the fish, or "o", for plotting lines with a dot at the head of the fish. Several values can be given.}

\item{schoolnumt}{is the number of time steps of the school, specified if the school data has been recycled and thus should be recycled when plotting as well.}

\item{add}{is TRUE if points are to be added to an existing plot.}

\item{aspect}{is used to set the dimension of the plotting frame (see aspect3d()).}
}
\value{

}
\description{
Plots fish as points or lines in 3-D, from data "psxf", "psyf", "pszf", "rtzf", "rtxf" located in the event.
}
\examples{
\dontrun{}

}
