% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.endcol.R
\name{get.endcol}
\alias{get.endcol}
\title{Adds color at the ends of the data.}
\usage{
get.endcol(endcol, col, rangevbsc, breaks, white)
}
\arguments{
\item{endcol}{The color to use for values above the highest break value.}

\item{col}{The present colors.}

\item{rangevbsc}{Range of the vbsc (sv) values.}

\item{breaks, white}{See \code{\link{cplot3d.event}}.}
}
\description{
Adds color at the ends of the data.
}
\examples{
\dontrun{}

}
