% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.xyzlim.R
\name{get.xyzlim}
\alias{get.xyzlim}
\title{Transforms 'xlim', 'ylim', and 'zlim' to a matrix of limits given the 'data' and the type of the inputs.}
\usage{
get.xyzlim(xlim, ylim, zlim, data = list(), t = 1, school = FALSE,
  cs.pos = "g", cs.xyzlim = "g")
}
\arguments{
\item{xlim}{is the x limits (x1, x2) of the plot. 'xlim' can be given as a list containing the elements 'xlim', 'ylim', and 'zlim' (names can be skipped, in which case the first element is interpreted as 'xlim', and so on). 'xlim' can also be given as a matrix cbind(xlim,ylim,zlim). 'xlim' can be used both to enlarge and to shrink the plotting frame (as opposed to plot3d(), in which xlim is only used to enlarge the plotting frame). If 'range' has been used in pplot3d.TSD(), xlim can be used to specify the limits presicely. Note that x1 > x2 is allowed and leads to a ‘reversed axis’. If one of xlim, ylim or zlim is NA, all the data and the entire sonar frame are included in the plot. The individual limits can take one of the following forms:}

\item{ylim}{is the same as 'xlim' but for the y-axis, and without the possibility of giving all of 'xlim', 'ylim', and 'zlim'.}

\item{zlim}{is the same as 'xlim' but for the z-axis, and without the possibility of giving all of 'xlim', 'ylim', and 'zlim'.}

\item{data}{is the list of TSD inputs as returned from read.event(var=c("vbsc","voxels","vessel")).}

\item{t}{is the time step used as input to 'xlim', 'ylim' and 'zlim', if these are functions.}

\item{school}{is TRUE if the school is to be plotted in pplot3d.TSD().}

\item{cs.pos}{indicates the coordinate system of the positions to be plotted, "g" representing the global coordinate system and "v" representing the coordinate system of the vessel.}

\item{cs.xyzlim}{indicates the coordinate system of the limits of the plot, "g" representing the global coordinate system and "v" representing the coordinate system of the vessel. cs.xyzlim="v" is particularly useful for setting the limits of the plotting frame to move with the vessel.}
}
\value{

}
\description{
Transforms 'xlim', 'ylim', and 'zlim' to a matrix of limits given the 'data' and the type of the inputs.
}
\examples{
\dontrun{}

}
