% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.colorsx.R
\name{heat.colorsx}
\alias{heat.colorsx}
\title{Returns a color vector of heat colors.}
\usage{
heat.colorsx(
  x,
  h = 1/6,
  v = 1,
  start = 0,
  end = 1,
  clamp = NULL,
  breakpoint = 0.75,
  flip = FALSE,
  alpha = 1,
  ...
)
}
\arguments{
\item{x}{is a numeric object to which rainbow colors combined with lightening and darkening should be matched.}

\item{h}{is the value hue at the end point of the heat color vector, defaulted to 1/6 to move from red to yellow.}

\item{v}{is the value (lightness) of the colors.}

\item{start}{and 'end' are the start and end colors, given as values in [0,1].}

\item{breakpoint}{is the breakpoint between 0 and 1 where the color vector turns at white.}

\item{flip}{is true if the color scale should be reversed.}

\item{alpha}{is the transparency.}

\item{...}{methods passed on to sub-functions (not used, but allowing for unused arguments).}
}
\description{
Returns a color vector of heat colors.
}
\examples{
\dontrun{
x=runif(1000)
pp(1,2)
plot(x, pch=".", cex=30, col=heat.colorsx(x))
plot(x, pch=".", cex=30, col=heat.colorsx(x, flip=TRUE))
}

}
