% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside.object3d.R
\name{inside.object3d}
\alias{inside.object3d}
\title{Returns a numeric or logical vector for whether the 3D positions given in 'x' are inside the object specified by 'object', 'center', 'angle' and 'par'.}
\usage{
inside.object3d(x, object = c("ellipsoid", "cuboid"), par = c(1, 1, 1),
  center = c(0, 0, 0), angle = 0, logical = FALSE)
}
\arguments{
\item{x}{is a three column matrix representing the x- y- and z-coordinates. If x is a vector, it is considered to be one single 3D point.}

\item{object}{is string representing the type of object to use. Currently implemented are "ellipsoid" and "cuboid" (may be abbreviated).}

\item{par}{is a vector of three elements representing the semi axis lengths for ellipsoid and the x-width, y-depth and z-height of cuboid.}

\item{center}{is a vector of three elements representing the centre position of the object.}

\item{angle}{is the angle of the major axis of the ellipsoid in the x-y-plane, in the case that object=="ellipsoid".}

\item{logical}{is TRUE if the output should be locigal instead of numeric.}
}
\value{

}
\description{
Returns a numeric or logical vector for whether the 3D positions given in 'x' are inside the object specified by 'object', 'center', 'angle' and 'par'.
}
\examples{
\dontrun{}

}
