% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locateNonVerticalBeamMode.R
\name{locateNonVerticalBeamMode}
\alias{locateNonVerticalBeamMode}
\title{Internal: Detects time steps with non-vertical beam mode, which can be used to override the beam mode specified in raw files.}
\usage{
locateNonVerticalBeamMode(t, event)
}
\value{

}
\description{
Internal: Detects time steps with non-vertical beam mode, which can be used to override the beam mode specified in raw files.
}
\examples{
\dontrun{}

}
