% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_ranges.R
\name{merge_ranges}
\alias{merge_ranges}
\title{Finds new ranges from matrices of ranges along the columns.}
\usage{
merge_ranges(...)
}
\arguments{
\item{...}{are matrices of range values along the columns. The matrices must have dimension [2 x p], where p is the number of ranges.}
}
\description{
Finds new ranges from matrices of ranges along the columns.
}
\examples{
\dontrun{}

}
