% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volume.ME70.R
\name{plot_volume.ME70}
\alias{plot_volume.ME70}
\title{A simple but unelegant way of plotting the sonar volume.}
\usage{
plot_volume.ME70(data, xyzlim = NULL, every = Inf, length_edges = 200,
  length_bows = 100, bowdens = 100, col = "black", ...)
}
\arguments{
\item{data}{is a list of data including the elements named "dira", "dire", "asps", "sint", "lenb", "rtzv", "psxv", "psyv", "pszv" and "psze".}

\item{xyzlim}{is a three column matrix of two rows, where the first column represents the plotting range of x-values and the second and third for the y- and z-values.}

\item{every}{is a single integer representing that the frame should be plotted every 'every' time step, where the first and last are included, or a vector of integers givint the time steps for which the frame should be plotted (not including the first and last time step).}

\item{length_edges}{is the number of points along the bows, which if chosen too low can result in lines appearing not to reach the bounding box.}

\item{length_bows}{is the number of points along the edges, which if chosen too low can result in lines appearing not to reach the bounding box.}

\item{bowdens}{is the distance in meters between bows.}

\item{col}{is the color for the plotted frame.}

\item{...}{are inputs passed on to lines3d().}
}
\value{

}
\description{
A simple but unelegant way of plotting the sonar volume.
}
\examples{
\dontrun{}

}
