% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volume.MS70.R
\name{plot_volume.MS70}
\alias{plot_volume.MS70}
\title{A simple but unelegant way of plotting the sonar volume.}
\usage{
plot_volume.MS70(data, xyzlim = NULL, t = 1, plot = c("proj",
  "frame"), cs.pos = "g", cs.view = "g", length_edges = 200,
  length_bows = 100, bowdens = 100, beamdens = 1, col = c("orange",
  "cornflowerblue"), sides = c("trbl", "trbl"), ends = FALSE, ...)
}
\arguments{
\item{data}{is a list of data including the elements named "dira", "dire", "asps", "sint", "lenb", "rtzv", "psxv", "psyv", "pszv" and "psze".}

\item{xyzlim}{is eihter a three column matrix of two rows, where the first column represents the plotting range of x-values and the second and third for the y- and z-values, or NA, indicating that the ranges of the sonar grid are to be returned.}

\item{t}{is the time step number of which to plot the sonar grid in the case that data$vbsc has a third dimension indicating time. 't' must be given as a single integer in the range seq_along(data$psxv).}

\item{plot}{is either "proj", for plotting the projection of the sonar volume onto the surface (the old method) or "frame" for plotting the frame around the sonar volume.}

\item{cs.pos}{is a single character representing the coordinate system used for the plotted data (one of "g" for global or "v" for vessel).}

\item{cs.view}{is a single character representing the coordinate system used for the view point (one of "g" for global or "v" for vessel).}

\item{length_edges}{is the number of points along the bows, which if chosen too low can result in lines appearing not to reach the bounding box.}

\item{length_bows}{is the number of points along the edges, which if chosen too low can result in lines appearing not to reach the bounding box.}

\item{bowdens}{is the distance in meters between bows. If given as a two element vector, the first is for the bows on the lower end of the sonar volume and the second is for the bows on the horizontal possibly projected end.}

\item{col}{is a vector of two color elements (string or integers) representing the color for the plotted frame, and the color of the lines forced onto the top of the bounding box respectively (in case parts of the frame is above the bounding box).}

\item{sides}{is a vector of two strings, one for the lines and one for the bows, including the characters "b" for bottom grid, "l" for left grid (as seed from the sonar), "t" for the top grid and "r" for the left grid.}

\item{...}{are inputs passed on to lines3d().}
}
\value{

}
\description{
A simple but unelegant way of plotting the sonar volume.
}
\examples{
\dontrun{}

}
