% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pplot3d.TSD.R
\name{pplot3d.TSD}
\alias{pplot3d.TSD}
\title{This function is the point representation analogue to cplot3d.TSD, and plots the acoustic data in 3d by point representation using the R package rgl (similar to applot.event() but only one time step) ????.}
\usage{
pplot3d.TSD(data, N = 1e+05, acca = NULL, fun = "mod",
  allert = 1e+08, nlim = NULL, fact = NULL, cols = c("black",
  "navyblue", "magenta4", "red3", "darkorange2", "yellow"), stretch = 1,
  scale.voxels = 1, rand.gen = c("unif", "beta", "norm"),
  beamstypes = 1, possample = 1, esnm = "MS70", var = c("vbsc",
  "sgsc", "pr0s", "sgs0", "sgsE", "sgsi", "sgsI", "psis", "tlns"),
  ind = list(-(1:150), NULL), range = list(), subset = NULL,
  ideal = TRUE, seabed = -12000, rot = 2, compensation = c("pitch",
  "roll"), plot = TRUE, cs.xyzlim = "g", add = FALSE, adds = NULL,
  xlim = NULL, ylim = NULL, zlim = NULL, view = c("free", "top",
  "bottom", "south", "west", "north", "east"), zoom = 0.7, fov = 60,
  school = FALSE, schoolcol = "purple", schoolsize = 0.3,
  schoolsample = 0.01, schoollen = 4, schoollwd = 1,
  schooltype = "p", schoolcrop = FALSE, plot.seg = FALSE,
  seg.col = "green", seg.alpha = 0.2, subdivide = 3,
  excl.neg = TRUE, object = c("ellipsoid", "cuboid"),
  par = double(3), center = c(0, 0, 0), angle = 0, aspect = "iso",
  nticks = 5, origin = 1, xlab = "x", ylab = "y", zlab = "z",
  full.box = FALSE, edge.vpos = "", line.vpos = 0, at.vpos = NULL,
  cex.vpos = 1, col.vpos = "blue", clock = "bbl", cex.clock = 1,
  adj.clock = c(0, 0),
  format.clock = "yyyy-mm-dd\\nHH:MM:SS.FFF\\nPing: indt",
  digits.clock = 2, lsp.clock = 0.04, col.clock = 4,
  sonar.grid = "frame", sonar.grid.col = c("orange", "cornflowerblue"),
  sonar.grid.lwd = 1, cs.pos = "g", cs.view = "g", sides = c("tb",
  "tb"), dens = c(200, 100, 100, 1), every = Inf,
  global.grid = FALSE, global.grid.lwd = 0.5, global.grid.lty = 1,
  ...)
}
\arguments{
\item{data}{is a list containing the data to plot, as returned from read.event(...,var=c("vbsc","voxels","ctd","beams","vessel","time")). May include school positions data$psxf, data$psyf and data$pszf. If non-school voxel probabilities are given in an array 'pr0s', or thresholded segmentation values 'sgsc'/'sgs0' are given, these can be plotted using the input 'var'.}
}
\value{

}
\description{
This function is the point representation analogue to cplot3d.TSD, and plots the acoustic data in 3d by point representation using the R package rgl (similar to applot.event() but only one time step) ????.
}
\examples{
\dontrun{}

}
