% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainbowx.R
\name{rainbowx}
\alias{rainbowx}
\title{Returns a color vector of rainbow colors. This is a special case of the combinedx() function.}
\usage{
rainbowx(x, s = 1, v = 1, start = 0, end = 1, clamp = NULL,
  flip = FALSE, alpha = 1, ...)
}
\arguments{
\item{x}{is a numeric object to which rainbow colors combined with lightening and darkening should be matched.}

\item{s}{is the range of the saturation values applied to the colors, where lower values suppress saturation. The default induces a clean rainbow.}

\item{v}{is the range of the value applied to the colors, where lower values causes darker colors. The default induces a clean rainbow.}

\item{start}{and 'end' are the start and end colors, given as values in [0,1].}

\item{flip}{is true if the color scale should be reversed.}

\item{alpha}{is the transparency.}

\item{...}{methods passed on to sub-functions (not used, but allowing for unused arguments).}
}
\value{

}
\description{
Returns a color vector of rainbow colors. This is a special case of the combinedx() function.
}
\examples{
\dontrun{
x=runif(1000)
pp(1,2)
plot(x, pch=".", cex=30, col=rainbowx(x))
plot(x, pch=".", cex=30, col=rainbowx(x, flip=TRUE))
}

}
