% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rseg.estimate.noise.R
\name{rseg.estimate.noise}
\alias{rseg.estimate.noise}
\title{Estimate the background (noise) level as specified by the Robust segmentation method in Profos.}
\usage{
rseg.estimate.noise(event, t_seq = 1, add = 2, t = 1, kern = 21L,
  type = 2, noisesmooth = 100, noise = NULL, ind = NULL)
}
\value{

}
\description{
Estimate the background (noise) level as specified by the Robust segmentation method in Profos.
}
\examples{
\dontrun{}

}
