% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rseg.event_block.R
\name{rseg.event_block}
\alias{rseg.event_block}
\title{Segments one ping with the (SX90) unbiased segmentatino method. Used in rseg.event().}
\usage{
rseg.event_block(t_seq, t, utim, noise, noisethr, thr, segfilesdir, sfnr,
  thersholdFromSchool, dBb1, dBb2, dBb3, dBb4, dBbs, nseg1, nseg2, Xcsz, add,
  event, kern, type, noisesmooth, ind, range, beamend, save.trh, save.data,
  save.ind, wt, esnm, cmpred, avoid, ellipsoid, nlmp, ...)
}
\arguments{
\item{t_seq}{is the time step index. One of 1:t.}

\item{t}{is the time steps to be segmented.}

\item{utim}{is a a vector of UNIX time points corresponding to 't'.}

\item{noisethr}{is the input dB above background threshold value.}

\item{segfilesdir}{is the directory of the segmentation files to be written.}

\item{sfnr}{is the segmentatino file number.}

\item{thersholdFromSchool}{is the (unbiased) threshold values defined from the singal.}

\item{dBb1, }{dBb2, dBb3, dBb4, dBbs  are the dB belov the signal, where the first four are parameters in the function f(S, SBR, Sv) with dBb1 being the constant term.}

\item{nseg1, }{nseg2  are the number of segmentations using backgroun-up and signal-down, respectively.}

\item{Xcsz}{is the circular size of the schools, calucated as 2*sqrt(A/pi).}

\item{kern}{is the standard deviation in units of the number of voxels used in a Gaussian kernel smoother along the beams (no smoothing if kern == 0 or length(kern) == 0, which is the default). If given as an integer, say 5L, median smoothing is applied instead of Gaussian.}

\item{type}{is the function to use across beams (one of "mean" and "median", where "mean" is used in PROFUS up to 2014, while "median" is more robust to the valus of the school).}

\item{save.trh}{is TRUE to save the threshold values used in the segmentation.}

\item{save.data}{is TRUE to save the data included in the segmentation mask.}

\item{wt}{is the number of time steps used as the width of the median filter along time steps.}

\item{esnm}{is the name of the acoustical instrument, given as a four character string. See sonR_implemented() for the implemented systems. May be given in 'data', and in lower case.}

\item{cmpred}{is a three column matrix holding the centers of mass of the ellipsoid outside which no samples are discarded from the segmentation mask. This matrix must have one row per time step, starting at the first time step.}

\item{ellipsoid}{is an optional vector of up to three elements specifying the semi axes of an ellipsoid centered at the center of mass of a user selected region, outside which voxels will not be included in the segmentation. The ellispoid can be set to move with the center of mass of the segmented voxels.}
}
\value{

}
\description{
Segments one ping with the (SX90) unbiased segmentatino method. Used in rseg.event().
}
\examples{
\dontrun{}

}
