% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rseg.event_oneping_write.R
\name{rseg.event_oneping_write}
\alias{rseg.event_oneping_write}
\title{Writes one ping of segmented data from the (SX90) unbiased segmentatino method. Used in rseg.event().}
\usage{
rseg.event_oneping_write(d, thisseg, thist_seq, t, utim, valid, subset, dimvbsc,
  segfilesdir, sfnr, dBan, dBb1, dBb2, dBb3, dBb4, dBbs, Xsbg, Xsb1, thisthr,
  thersholdFromSchool, Xcsz, save.trh, save.data, save.ind, kern, wt, type,
  cmpred, ellipsoid, file_thist_seq, reserve)
}
\arguments{
\item{d}{is the list of data including segmentation and acoustic data.}

\item{thisseg}{is the segmentatino mask.}

\item{thist_seq}{is the time step index. One of 1:t.}

\item{t}{is all the time steps to be segmented.}

\item{utim}{is a a vector of UNIX time points corresponding to 't'.}

\item{valid}{is indices of the valid region of the sonar volume.}

\item{subset}{is the subset related to excusion of vessel realted noise such as wake.}

\item{dimvbsc}{are the dimenstions of the acoustic data.}

\item{segfilesdir}{is the directory of the segmentation files to be written.}

\item{sfnr}{is the segmentatino file number.}

\item{dBan}{is the dB above background.}

\item{dBb1, }{dBb2, dBb3, dBb4, dBbs  are the dB belov the signal, where the first four are parameters in the function f(S, SBR, Sv) with dBb1 being the constant term.}

\item{Xsbg, }{Xsb1  are vectors of the smoothed background estimate (noise is the median over consecutive pings and Gaussian kernel smoothed along beams, whereas Xsb1 is not).}

\item{thisthr}{is a vector of the above background-thresholds.}

\item{thersholdFromSchool}{is the (unbiased) threshold values defined from the singal.}

\item{Xcsz}{is the circular size of the schools, calucated as 2*sqrt(A/pi).}

\item{save.trh}{is TRUE to save the threshold values used in the segmentation.}

\item{save.data}{is TRUE to save the data included in the segmentation mask.}

\item{kern}{is the standard deviation in units of the number of voxels used in a Gaussian kernel smoother along the beams (no smoothing if kern == 0 or length(kern) == 0, which is the default). If given as an integer, say 5L, median smoothing is applied instead of Gaussian.}

\item{wt}{is the number of time steps used as the width of the median filter along time steps.}

\item{type}{is the function to use across beams (one of "mean" and "median", where "mean" is used in PROFUS up to 2014, while "median" is more robust to the valus of the school).}

\item{cmpred}{is a three column matrix holding the centers of mass of the ellipsoid outside which no samples are discarded from the segmentation mask. This matrix must have one row per time step, starting at the first time step.}

\item{ellipsoid}{is an optional vector of up to three elements specifying the semi axes of an ellipsoid centered at the center of mass of a user selected region, outside which voxels will not be included in the segmentation. The ellispoid can be set to move with the center of mass of the segmented voxels.}

\item{file_thist_seq}{is the time index in the block. If this is equal to 'thist_seq', a new file will be written. Otherwise data will be appended to the file with name given by 'file_thist_seq'.}

\item{reserve}{is used in write.TSD() to reserve time steps for the whole block of time steps.}
}
\value{

}
\description{
Writes one ping of segmented data from the (SX90) unbiased segmentatino method. Used in rseg.event().
}
\examples{
\dontrun{}

}
