% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.view3d.R
\name{set.view3d}
\alias{set.view3d}
\title{Function for extracting the view point of an rgl plot.}
\usage{
set.view3d(view = c("free", "top", "bottom", "south", "west", "north",
  "east"), cs.pos = "g", cs.view = "g", rtzv = 0)
}
\arguments{
\item{view}{defines the view point of the user and has 4 possible values:}

\item{cs.pos}{defines the coordinate system to use when generating the points. Given as one of "g" or "v" representing the global coordinate and the coordinate system of the vessel. If given as a three element vector, 'cs.pos' is interpreted as c(cs.pos,cs.view).}

\item{cs.view}{defines the coordinate system to use when specifying the view point. Given as one of "g" or "v" representing the global coordinate and the coordinate system of the vessel.}

\item{rtzv}{is the heading of the vessel.}
}
\value{

}
\description{
Function for extracting the view point of an rgl plot.
}
\examples{
\dontrun{}

}
