% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv2pos_get_ind.R
\name{sv2pos_get_ind}
\alias{sv2pos_get_ind}
\title{Extracts a subset of TSD data according to the array subset 'ind' and/or the cartesian subset 'range' and/or the logical/numeric vector of subscripts 'subset'.}
\usage{
sv2pos_get_ind(
  data,
  ind = list(),
  range = list(),
  subset = NULL,
  strict = TRUE
)
}
\arguments{
\item{data}{is a list of elements named according to the TSD file format.}

\item{ind}{is a list of indexes, as typed into the [] of an array, where 0 and NULL denotes all indexes.}

\item{range}{is a list of elements with names matching names if 'data', specifying the range of the corresponding elements.}

\item{subset}{is a numeric or logical vector/expression indicating elements or rows to keep (as used in subset()). Missing values are taken as false, and subset=0 or subset=NULL indicates no subsetting.}

\item{strict}{is TRUE if strict inequality is to be used when subsetting according to 'range'.}
}
\description{
Extracts a subset of TSD data according to the array subset 'ind' and/or the cartesian subset 'range' and/or the logical/numeric vector of subscripts 'subset'.
}
\examples{
\dontrun{}

}
