% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SX90_nr0a.R
\name{SX90_nr0a}
\alias{SX90_nr0a}
\title{Generates close range noise in active mode to be used for fishery sonar. The noise is generated in non-TVG-amplified mode such that when TVG-amplified, the noise follows a pure log r-relationship, i.e., m * log(r). The term "close range noise" is here used somewhat misleading, since it is merely used to mimic the background level of a fishery sonar, which includes surface reverberation and possibly other backscattering which does not fall into the definition of noise in this system (all energy not backscattered from objects).}
\usage{
SX90_nr0a(beams, SNref = -65, rref = 600, SN0 = -25, rjoin = 20,
  m = 2, TVG.exp = 2)
}
\arguments{
\item{beams}{is a list of beam configuration data, specifically}

\item{SNref}{is the reference noise level in dB at range 'rref'.}

\item{rref}{is the range of the reference noise level.}

\item{SN0}{is the noise level at zero range.}

\item{rjoin}{is the range to the point where the transient like close range noise (passive mode) joins the active mode close range noise.}

\item{m}{specifies the TVG shape of the noise as modeled by 10 m log(r). If m=2, the noise (except absorption) follows the clean background noise, whereas if m=0, the noise (except absorption) is flat (independen of the range form the sonar). The absorption is always present, causing decreasing signal to noise ratio (SNR).}

\item{TVG.exp}{a vector of two elements specifying the near field of the noise, where the first element is the noise}
}
\value{

}
\description{
Generates close range noise in active mode to be used for fishery sonar. The noise is generated in non-TVG-amplified mode such that when TVG-amplified, the noise follows a pure log r-relationship, i.e., m * log(r). The term "close range noise" is here used somewhat misleading, since it is merely used to mimic the background level of a fishery sonar, which includes surface reverberation and possibly other backscattering which does not fall into the definition of noise in this system (all energy not backscattered from objects).
}
\examples{
\dontrun{}

}
