% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.colorsx.R
\name{terrain.colorsx}
\alias{terrain.colorsx}
\title{Returns a color vector of terrain colors.}
\usage{
terrain.colorsx(x, start = 0, end = 1, clamp = NULL, breakpoint = 0.5,
  flip = FALSE, alpha = 1, ...)
}
\arguments{
\item{x}{is a numeric object to which rainbow colors combined with lightening and darkening should be matched.}

\item{start}{and 'end' are the start and end colors, given as values in [0,1].}

\item{breakpoint}{is the breakpoint between 0 and 1 where the color vector turns at white.}

\item{flip}{is true if the color scale should be reversed.}

\item{alpha}{is the transparency.}

\item{...}{methods passed on to sub-functions (not used, but allowing for unused arguments).}
}
\value{

}
\description{
Returns a color vector of terrain colors.
}
\examples{
\dontrun{
x=runif(1000)
pp(1,2)
plot(x,pch=".",cex=30,col=terrain.colorsx(x))
plot(x,pch=".",cex=30,col=terrain.colorsx(x,flip=TRUE))
}

}
