% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo.colorsx.R
\name{topo.colorsx}
\alias{topo.colorsx}
\title{Returns a color vector of topographic colors.}
\usage{
topo.colorsx(x, v = 1, start = 0, end = 1, clamp = NULL,
  breakpoint = c(1/3, 2/3), flip = FALSE, alpha = 1, ...)
}
\arguments{
\item{x}{is a numeric object to which rainbow colors combined with lightening and darkening should be matched.}

\item{v}{is the value (lightness) of the colors.}

\item{start}{and 'end' are the start and end colors, given as values in [0,1].}

\item{breakpoint}{is the vector of two breakpoints between 0 and 1 where the color vector turns from blue to green and from green to yellow.}

\item{flip}{is true if the color scale should be reversed.}

\item{alpha}{is the transparency.}

\item{...}{methods passed on to sub-functions (not used, but allowing for unused arguments).}
}
\value{

}
\description{
Returns a color vector of topographic colors.
}
\examples{
\dontrun{
x=runif(1000)
pp(1,2)
plot(x,pch=".",cex=30,col=topo.colorsx(x))
plot(x,pch=".",cex=30,col=topo.colorsx(x,flip=TRUE))
}

}
